Genre handlers.

+++ Backend/Movieverse.Application/Common/PaginatedList.cs
@@ -6,15 +6,15 @@ namespace Movieverse.Application.Common;
 public sealed class PaginatedList<TKey> : IPaginatedList<TKey>
 {
 	public List<TKey> Items { get; init; } = null!;
-	public int? PageNumber { get; init; }
-	public int? TotalPages { get; init; }
-	public int TotalCount { get; init; }
+	public short? PageNumber { get; init; }
+	public short? TotalPages { get; init; }
+	public short TotalCount { get; init; }
 	public bool HasPreviousPage => PageNumber > 1;
 	public bool HasNextPage => PageNumber < TotalPages;
 
-	public static async Task<PaginatedList<TKey>> CreateAsync(IQueryable<TKey> source, int? pageNumber, int? pageSize)
+	public static async Task<PaginatedList<TKey>> CreateAsync(IQueryable<TKey> source, short? pageNumber, short? pageSize)
 	{
-		var count = await source.CountAsync();
+		var count = (short)await source.CountAsync();
 
 		if (pageSize.HasValue && pageNumber.HasValue)
 		{
@@ -27,15 +27,15 @@ public sealed class PaginatedList<TKey> : IPaginatedList<TKey>
 		{
 			Items = items,
 			PageNumber = pageNumber,
-			TotalPages = pageSize == null ? null : (int)Math.Ceiling(count / (double)pageSize.Value),
+			TotalPages = pageSize == null ? null : (short)Math.Ceiling(count / (double)pageSize.Value),
 			TotalCount = count
 		};
 	}
     
-	public static PaginatedList<TKey> Create(IEnumerable<TKey> source, int? pageNumber, int? pageSize)
+	public static PaginatedList<TKey> Create(IEnumerable<TKey> source, short? pageNumber, short? pageSize)
 	{
 		var temp = source.ToList();
-		var count = temp.Count;
+		var count = (short)temp.Count;
         
 		if (pageSize.HasValue && pageNumber.HasValue)
 		{
@@ -46,7 +46,7 @@ public sealed class PaginatedList<TKey> : IPaginatedList<TKey>
 		{
 			Items = temp,
 			PageNumber = pageNumber,
-			TotalPages = pageSize == null ? null : (int)Math.Ceiling(count / (double)pageSize.Value),
+			TotalPages = pageSize == null ? null : (short)Math.Ceiling(count / (double)pageSize.Value),
 			TotalCount = count
 		};
 	}
@@ -55,12 +55,12 @@ public sealed class PaginatedList<TKey> : IPaginatedList<TKey>
 public static class PaginatedListExtensions
 {
 	public static Task<PaginatedList<TDestination>> ToPaginatedListAsync<TDestination>(this IQueryable<TDestination> queryable,
-		int? pageNumber, int? pageSize)
+		short? pageNumber, short? pageSize)
 		where TDestination : class
 		=> PaginatedList<TDestination>.CreateAsync(queryable, pageNumber, pageSize);
 
 	public static PaginatedList<TDestination> ToPaginatedList<TDestination>(this IEnumerable<TDestination> enumerable,
-		int? pageNumber, int? pageSize)
+		short? pageNumber, short? pageSize)
 		where TDestination : class
 		=> PaginatedList<TDestination>.Create(enumerable, pageNumber, pageSize);
 }
