+++ Backend/Movieverse.Domain/Common/IPaginatedList.cs
@@ -3,9 +3,9 @@
 public interface IPaginatedList<TKey>
 {
 	List<TKey> Items { get; init; }
-	int? PageNumber { get; init; }
-	int? TotalPages { get; init; }
-	int TotalCount { get; init; }
+	short? PageNumber { get; init; }
+	short? TotalPages { get; init; }
+	short TotalCount { get; init; }
 	bool HasPreviousPage => PageNumber > 1;
 	bool HasNextPage => PageNumber < TotalPages;
 }
